Sat, Feb 25, 2023, 9:49 AM +06:00

+++ test/learn.c
@@ -2,14 +2,14 @@
 
 int main()
 {
-    long long int price = 250, paid, change, unit; // 1 unit = 2+1 pizza
+    double price = 250, paid, change, unit; // 1 unit = 2+1 pizza
     while(1)
     {
         change = 0;
         unit = 0;
         paid = 0;
         printf("\nEnter the ammount you paid: ");
-        scanf("%lld", &paid);
+        scanf("%lf", &paid);
         if(paid == EOF){break;}
 
         if(paid > 250)
@@ -22,7 +22,7 @@ int main()
             else if(paid >= 500)
             {
                 unit = paid / 500;
-                printf("You are getting %lld pizza's, sir!\n", (unit*3));
+                printf("You are getting %.0lf pizza's, sir!\n", (unit*3));
                 price = unit * 500;
                 change = paid - price;
             }
@@ -33,7 +33,7 @@ int main()
             change = paid;
         }
 
-        printf("Your Change is: %lld$\n", change);
+        printf("Your Change is: %.0lf$\n", change);
 
     }
     return 0;
